refactor(core): move gitignore helper to its own module
